requirements.txt
fastapi
uvicorn
jsonschema
python-docx
docxtpl
pytest
httpx
python-multipart